+++ services/__tests__/static.test.js
@@ -184,7 +184,7 @@ describe('getStaticImage', () => {
           }
         }
       ],
-      insertOverlayBeforeLayer: 'national_park'
+      before_layer: 'national_park'
     });
     expect(tu.requestConfig(service)).toEqual({
       method: 'GET',
@@ -415,5 +415,222 @@ describe('getStaticImage', () => {
         ]
       });
     }).toThrow(/fillOpacity requires fillColor/);
+  });
+
+  test('setfilter', () => {
+    service.getStaticImage({
+      ownerId: 'mapbox',
+      styleId: 'streets-v10',
+      width: 200,
+      height: 300,
+      position: {
+        coordinates: [12, 13],
+        zoom: 3
+      },
+      setfilter: ['>', 'height', 300],
+      layer_id: 'building'
+    });
+    expect(tu.requestConfig(service)).toEqual({
+      method: 'GET',
+      path: '/styles/v1/:ownerId/:styleId/static/12,13,3/200x300',
+      query: {
+        setfilter: '[">","height",300]',
+        layer_id: 'building'
+      },
+      params: { ownerId: 'mapbox', styleId: 'streets-v10' }
+    });
+  });
+
+  test('Must include layer_id in setfilter request', () => {
+    expect(() => {
+      service.getStaticImage({
+        ownerId: 'mapbox',
+        styleId: 'streets-v10',
+        width: 200,
+        height: 300,
+        position: {
+          coordinates: [12, 13],
+          zoom: 3
+        },
+        setfilter: ['in', 'code', 'CA']
+      });
+    }).toThrow(/Must include layer_id in setfilter request/);
+  });
+
+  test('Auto extent cannot be used with style parameters and no overlay', () => {
+    expect(() => {
+      service.getStaticImage({
+        ownerId: 'mapbox',
+        styleId: 'streets-v10',
+        width: 200,
+        height: 300,
+        position: 'auto',
+        setfilter: ['in', 'code', 'CA'],
+        layer_id: 'tunnel-street-minor-low'
+      });
+    }).toThrow(
+      /Auto extent cannot be used with style parameters and no overlay/
+    );
+  });
+
+  test('Auto extent can be used with setfilter when overlays are defined', () => {
+    service.getStaticImage({
+      ownerId: 'mapbox',
+      styleId: 'streets-v10',
+      width: 200,
+      height: 300,
+      position: 'auto',
+      overlays: [
+        {
+          marker: {
+            coordinates: [12.2, 12.8]
+          }
+        }
+      ],
+      setfilter: ['in', 'code', 'CA'],
+      layer_id: 'tunnel-street-minor-low'
+    });
+    expect(tu.requestConfig(service)).toEqual({
+      method: 'GET',
+      path: '/styles/v1/:ownerId/:styleId/static/pin-s(12.2,12.8)/auto/200x300',
+      query: {
+        setfilter: '["in","code","CA"]',
+        layer_id: 'tunnel-street-minor-low'
+      },
+      params: { ownerId: 'mapbox', styleId: 'streets-v10' }
+    });
+  });
+
+  test('addlayer', () => {
+    service.getStaticImage({
+      ownerId: 'mapbox',
+      styleId: 'streets-v10',
+      width: 200,
+      height: 300,
+      position: {
+        coordinates: [12, 13],
+        zoom: 3
+      },
+      addlayer: {
+        id: 'tall-buildings',
+        type: 'fill',
+        source: 'composite',
+        'source-layer': 'building',
+        filter: [
+          'all',
+          ['>=', ['get', 'height'], 150],
+          ['match', ['get', 'underground'], ['false'], true, false]
+        ],
+        paint: { 'fill-color': '%235E8DFF', 'fill-opacity': 0.5 }
+      },
+      before_layer: 'tunnel-street-minor-low'
+    });
+    expect(tu.requestConfig(service)).toEqual({
+      method: 'GET',
+      path: '/styles/v1/:ownerId/:styleId/static/12,13,3/200x300',
+      query: {
+        addlayer:
+          '{"id":"tall-buildings","type":"fill","source":"composite","source-layer":"building","filter":["all",[">=",["get","height"],150],["match",["get","underground"],["false"],true,false]],"paint":{"fill-color":"%235E8DFF","fill-opacity":0.5}}',
+        before_layer: 'tunnel-street-minor-low'
+      },
+      params: { ownerId: 'mapbox', styleId: 'streets-v10' }
+    });
+  });
+
+  test('Auto extent cannot be used with style parameters and no overlay', () => {
+    expect(() => {
+      service.getStaticImage({
+        ownerId: 'mapbox',
+        styleId: 'streets-v10',
+        width: 200,
+        height: 300,
+        position: 'auto',
+        addlayer: {
+          id: 'tall-buildings',
+          type: 'fill',
+          source: 'composite',
+          'source-layer': 'building',
+          filter: [
+            'all',
+            ['>=', ['get', 'height'], 150],
+            ['match', ['get', 'underground'], ['false'], true, false]
+          ],
+          paint: { 'fill-color': '%235E8DFF', 'fill-opacity': 0.5 }
+        },
+        before_layer: 'tunnel-street-minor-low'
+      });
+    }).toThrow(
+      /Auto extent cannot be used with style parameters and no overlay/
+    );
+  });
+
+  test('Auto extent can be used with addlayer when overlays are defined', () => {
+    service.getStaticImage({
+      ownerId: 'mapbox',
+      styleId: 'streets-v10',
+      width: 200,
+      height: 300,
+      position: 'auto',
+      overlays: [
+        {
+          marker: {
+            coordinates: [12.2, 12.8]
+          }
+        }
+      ],
+      addlayer: {
+        id: 'tall-buildings',
+        type: 'fill',
+        source: 'composite',
+        'source-layer': 'building',
+        filter: [
+          'all',
+          ['>=', ['get', 'height'], 150],
+          ['match', ['get', 'underground'], ['false'], true, false]
+        ],
+        paint: { 'fill-color': '%235E8DFF', 'fill-opacity': 0.5 }
+      },
+      before_layer: 'tunnel-street-minor-low'
+    });
+    expect(tu.requestConfig(service)).toEqual({
+      method: 'GET',
+      path: '/styles/v1/:ownerId/:styleId/static/pin-s(12.2,12.8)/auto/200x300',
+      query: {
+        addlayer:
+          '{"id":"tall-buildings","type":"fill","source":"composite","source-layer":"building","filter":["all",[">=",["get","height"],150],["match",["get","underground"],["false"],true,false]],"paint":{"fill-color":"%235E8DFF","fill-opacity":0.5}}',
+        before_layer: 'tunnel-street-minor-low'
+      },
+      params: { ownerId: 'mapbox', styleId: 'streets-v10' }
+    });
+  });
+
+  test('addlayer and setfilter cannot be used in the same request', () => {
+    expect(() => {
+      service.getStaticImage({
+        ownerId: 'mapbox',
+        styleId: 'streets-v10',
+        width: 200,
+        height: 300,
+        position: {
+          coordinates: [12, 13],
+          zoom: 3
+        },
+        addlayer: {
+          id: 'tall-buildings',
+          type: 'fill',
+          source: 'composite',
+          'source-layer': 'building',
+          filter: [
+            'all',
+            ['>=', ['get', 'height'], 150],
+            ['match', ['get', 'underground'], ['false'], true, false]
+          ],
+          paint: { 'fill-color': '%235E8DFF', 'fill-opacity': 0.5 }
+        },
+        before_layer: 'tunnel-street-minor-low',
+        setfilter: ['in', 'code', 'CA'],
+        layer_id: 'tunnel-street-minor-low'
+      });
+    }).toThrow(/addlayer and setfilter cannot be used in the same request/);
   });
 });